restructured file system to represent component organization
