--- debug.log
@@ -1,1 +1,0 @@
-[1013/175146.966:ERROR:directory_reader_win.cc(43)] FindFirstFile: The system cannot find the path specified. (0x3)
